Change prognum, versnum, minvers, and maxvers in progpin to u_int32_t from u_long
This fixes a problem where "rpcinfo -T transport host prognum"
fails on a PPC64 because CLNT_CONTROL expects the version
number to be a 32 bit quantity.  u_long probably works fine
on little endian machines, but won't work on big endian machines.

+++ src/rpcinfo.c
@@ -1591,7 +1591,7 @@ progping (netid, argc, argv)
   CLIENT *client;
   struct timeval to;
   enum clnt_stat rpc_stat;
-  u_long prognum, versnum, minvers, maxvers;
+  u_int32_t prognum, versnum, minvers, maxvers;
   struct rpc_err rpcerr;
   int failure = 0;
   struct netconfig *nconf;
